VERSION 1:
- Installing Moonsharp basics
- - 1 Hour
- Setting up PlayerScript
- - 2 Hours
- Creating simple editor/scene
- - 3.5 Hours
- Implementing Proxies
- - 4 Hours

VERSION 2 (current):

Setup Scene (Sprite)
- 2 Hours

Setup basic Robot
- 1 Hour

Setup Movement
- 1 Hour

Selection of tiles
- 1 Hour

Robot Classes
- 1 Hour

+++ Assets/Pieecs/Scripts/PlayerObject.cs
@@ -6,13 +6,13 @@ public abstract class PlayerObject : MonoBehaviour
 {
 
 
-    public void Setup(Player player, int Health)
+    public void Setup(Player player, float Health)
     {
         this.Player = player;
         this.Health = Health;
     }
 
-    public void RemoveHealth(int damage)
+    public void RemoveHealth(float damage)
     {
         this.Health -= damage;
         if (this.Health <= 0)
@@ -33,7 +33,7 @@ public abstract class PlayerObject : MonoBehaviour
     
 
 
-    public int Health { get; protected set; }
+    public float Health { get; protected set; }
